% Generated by roxygen2 (4.0.1): do not edit by hand
\name{all_undp_tables}
\alias{all_undp_tables}
\title{Get all tables from the API}
\usage{
all_undp_tables()
}
\value{
list of dataframes
}
\description{
This function downloads all available tables from the UNdp API
}
\details{
Failed downloads return an error string
}

