% Generated by roxygen2 (4.0.1): do not edit by hand
\name{content_to_df}
\alias{content_to_df}
\title{Converts to data frame even with missing columns}
\usage{
content_to_df(con)
}
\arguments{
\item{con}{a list as output by content(response)}
}
\value{
dataframe
}
\description{
If all items are of the same length, just goes ahead and converts to df
If the items are of different lengths, assume the longest has all the columns,
fill in the gaps with NA in the other columns and return in the original column order
}
\author{
David A Springate \email{daspringate@gmail.com}
}

