% Generated by roxygen2 (4.0.1): do not edit by hand
\name{posixify}
\alias{posixify}
\title{Convert Socrata calendar_date string to POSIX}
\usage{
posixify(x)
}
\arguments{
\item{x}{char in Socrata calendar_date format}
}
\value{
a POSIX date
}
\description{
Convert Socrata calendar_date string to POSIX
}
\author{
Hugh J. Devlin, Ph. D. \email{Hugh.Devlin@cityofchicago.org}
}

