% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{rUNdp}
\alias{rUNdp}
\alias{rUNdp-package}
\title{rUNdp, R integration with the UN development programme API
Copyright (C) 2014 David A Springate}
\description{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License, version 2,
as published by the Free Software Foundation.
}
\details{
rClinicalCodes is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License along
with this program; if not, write to the Free Software Foundation, Inc.,
51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}

