% Generated by roxygen2 (4.0.1): do not edit by hand
\name{validateUrl}
\alias{validateUrl}
\title{Convert, if necessary, URL to valid REST API URL supported by Socrata.}
\usage{
validateUrl(url)
}
\arguments{
\item{url}{a string; character vector of length one}
}
\value{
a valid Url
}
\description{
Will convert a human-readable URL to a valid REST API call
supported by Socrata. It will accept a valid API URL if provided
by users and will also convert a human-readable URL to a valid API
URL.
}
\author{
Tom Schenk Jr \email{tom.schenk@cityofchicago.org}
}

